moved PlotDepthSlice
