+++ mtpy/modeling/ModEM/__init__.py
@@ -10,9 +10,10 @@ from .config import ModEMConfig
 from .model_manipulator import ModelManipulator
 from .plot_response import PlotResponse
 from .plot_pt_maps import PlotPTMaps
+from .plot_depth_slice import PlotDepthSlice
 
 __all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model', 'Residual',
            'ControlInv', 'ControlFwd', 'Covariance', 'ModEMConfig', 'ModelManipulator',
-           'PlotResponse', 'PlotPTMaps']
+           'PlotResponse', 'PlotPTMaps', 'PlotDepthSlice']
